Extract draw_circle_filled method
