+++ pupil_src/shared_modules/gl_utils/__init__.py
@@ -8,6 +8,7 @@ Lesser General Public License (LGPL v3.0).
 See COPYING and COPYING.LESSER for license details.
 ---------------------------------------------------------------------------~(*)
 """
+from .draw import draw_circle_filled
 from .trackball import Trackball
 from .utils import (
     _Rectangle,
